requirements.txt
amqp==5.1.1
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
Babel==2.10.3
billiard==3.6.4.0
celery==5.2.3
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
Deprecated==1.2.13
Django==4.0.1
django-environ==0.8.1
django-filter==22.1
django-money==3.0.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==4.8.0
flake8==4.0.1
flower==1.1.0
humanize==4.2.3
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==21.3
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
py-moneyed==2.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.9
pytest==7.1.2
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-django==4.5.2
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
pytz==2022.1
redis==4.1.2
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
gunicorn==20.0.4
python-dotenv==0.19.2
django-countries==7.5.1
drf-yasg==1.21.4
django-debug-toolbar==3.8.1
drf-nested-routers==0.93.4
pre-commit==3.0.4
pylint==2.16.1
django-rest-authemail==2.1.4
authemail==2.1.4